+++ templates/index.html
@@ -13,7 +13,7 @@
     <form action="/course_select" method="post">
       <label for="fdept">Course Department:</label>
         <select name="fdept" id="fdept">
-            <option value="cst">CST</option>
+            <option value="cst">CIS</option>
             <option value="arch">ARCH</option>
             <option value="acct">ACCT</option>
             <option value="bioe">BIOE</option>
@@ -30,7 +30,7 @@
             <option value="fma">FMA</option>
         </select><br>
       <label for = "fcrs1">1st Course Number: </label>
-        <input type="text" id="fcrs1" name="fcrs1" value="3219"><br>
+        <input type="text" id="fcrs1" name="fcrs1" value=""><br>
 
       <label for = "fcrs2">2nd Course Number: </label>
         <input type="text" id="fcrs2" name="fcrs2" value=""><br>
